Build: Split packages and blocks to their webpack configs
